Separación en un archivo independiente las funciones de estadisticas
